refactor: consolidate duplicate identite_pg middleware implementations
- Merge identite_pg.ts functionality into set_identite_pg.ts
- Add set_identite_pg_database() function for connection string setup
- Remove duplicate files and update all imports
- Follow established set_ naming convention for middleware
- Fix import references and TypeScript errors
- Add proper typing for destructured parameters
