+++ packages/~/app/api/src/readyz/index.ts
@@ -1,7 +1,7 @@
 //
 
 import env from "@~/app.core/config";
-import { identite_pg_database } from "@~/app.middleware/identite_pg";
+import { set_identite_pg_database } from "@~/app.middleware/set_identite_pg";
 import { get_zammad_me } from "@~/zammad.lib";
 import { to } from "await-to-js";
 import { sql } from "drizzle-orm";
@@ -26,7 +26,7 @@ export default new Hono()
   })
   .get(
     "/drizzle/identite",
-    identite_pg_database({ connectionString: env.DATABASE_URL }),
+    set_identite_pg_database({ connectionString: env.DATABASE_URL }),
     async ({ text, var: { identite_pg } }) => {
       const [, is_ok] = await to(identite_pg.execute(sql`SELECT 1`));
       return text("[+]drizzle identite connection " + (is_ok ? "OK" : "FAIL"));
